requirements.txt
PySide6
cryptography
mfrc522
spidev
RPi.GPIO
python-dotenv